+++ data/fetcher.go
@@ -1,5 +1,6 @@
 package data
 
+// Fetcher is the common interface of a fetcher service.
 type Fetcher interface {
 	Run() error
 	Stop() error