+++ data/fetcher/http_runner/runner_test.go
@@ -75,10 +75,6 @@ func TestHttpRunner(t *testing.T) {
 			path: "btick",
 		},
 		{
-			ch:   runner.GetTradeHistoryTicker(),
-			path: "ttick",
-		},
-		{
 			ch:   runner.GetGlobalDataTicker(),
 			path: "gtick",
 		},