+++ exchange/bittrex_storage.go
@@ -4,7 +4,7 @@ import (
 	"github.com/KyberNetwork/reserve-data/common"
 )
 
-// This storage is used to store deposit histories
+// BittrexStorage is used to store deposit histories
 // to check if a deposit history id is registered for
 // a different deposit already
 type BittrexStorage interface {
@@ -13,5 +13,4 @@ type BittrexStorage interface {
 	StoreTradeHistory(data common.ExchangeTradeHistory) error
 
 	GetTradeHistory(fromTime, toTime uint64) (common.ExchangeTradeHistory, error)
-	GetLastIDTradeHistory(exchange, pair string) (string, error)
 }