+++ exchange/huobi_blockchain.go
@@ -7,6 +7,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 )
 
+// HuobiBlockchain contains methods to interact with blockchain from Huobi address.
 type HuobiBlockchain interface {
 	SendETHFromAccountToExchange(amount *big.Int, exchangeAddress ethereum.Address) (*types.Transaction, error)
 	SendTokenFromAccountToExchange(amount *big.Int, exchangeAddress ethereum.Address, tokenAddress ethereum.Address) (*types.Transaction, error)