+++ stat/fee_setrate_storage.go
@@ -4,6 +4,7 @@ import (
 	"github.com/KyberNetwork/reserve-data/common"
 )
 
+// FeeSetRateStorage it the storage interface to keep track of SetRate fee.
 type FeeSetRateStorage interface {
 	GetLastBlockChecked() (uint64, error)
 	StoreTransaction(tx []common.SetRateTxInfo) error